added new level indicators

+++ core/src/com/mygdx/game2/Animation.java
@@ -8,8 +8,8 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 public abstract class Animation {
     //animations are things that happen in the environment but interact with literally nothing, particles interact with things
     // but they dont change them, animations literally dont interact they just do their animation
-    public int x = 0;
-    public int y = 0;
+    public float x = 0;
+    public float y = 0;
     public float velX = 0;
     public float velY = 0;
     public boolean destroyed;
